+++ app/components/app/chat/EditBubble.vue
@@ -13,6 +13,7 @@ const translation = useTranslation({
   chatUsername: username,
   toNative: false,
   refetchOnTranslate: true,
+  messageIsBot: false,
 })
 
 watch(translation.isLoading, (value) => {